asserv + strategie

+++ Informatique/robot-teensy/src/Odometry.h
@@ -1,9 +1,10 @@
 #include <Arduino.h>
 #include <math.h>
 #include "Encoder.h"
+#include "Strategy.h"
 
 // Variables
-extern float g_x, g_y, g_angle;
+extern double g_x, g_y, g_angle;
 
 // Prototypes
 uint8_t Odometry_Init(void);
